+++ backend/SmartPlaylist/Contracts/SmartPlaylistNewItemOrderDto.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace SmartPlaylist.Contracts
+{
+    [Serializable]
+    public class SmartPlaylistNewItemOrderDto
+    {
+        public bool HasSort { get; set; }
+        public string OrderBy { get; set; }
+    }
+}