+++ frontend/src/common/components/Toggle.tsx
@@ -4,7 +4,10 @@ import './Toggle.css';
 export type ToggleProps = {
     id: string,
     checked: boolean,
-    onChange(change: boolean): void,
+    label?: string,
+    switchStyle?: React.CSSProperties,
+    containerStyle?: React.CSSProperties,
+    onToggled(change: boolean): void,
 
 } & BaseProps;
 
@@ -13,29 +16,33 @@ export const Toggle: React.FC<ToggleProps> = props => {
         if (e.keyCode !== 32) return;
 
         e.preventDefault();
-        props.onChange(!props.checked)
+        props.onToggled(!props.checked)
     }
 
-    return (
-        <div className="toggle-container">
-            <div className="toggle-switch">
-                <input
-                    type="checkbox"
-                    className="toggle-checkbox"
-                    id={props.id}
-                    name={props.id}
-                    checked={props.checked}
-                    onKeyDown={(e) => { handleKeyPress(e) }}
-                    onChange={e => {
-                        console.log('change');
-                        props.onChange(e.target.checked);
-                    }}
-                />
-                <label className="toggle-label" htmlFor={props.id}>
-                    <span className="toggle-inner" />
-                    <span className="toggle-toggle-switch" />
-                </label>
-            </div>
+    const main = <div className="toggle-container" style={props.containerStyle}>
+        <div className="toggle-switch">
+            <input
+                type="checkbox"
+                className="toggle-checkbox"
+                id={props.id}
+                name={props.id}
+                checked={props.checked}
+                onKeyDown={(e) => { handleKeyPress(e) }}
+                onChange={e => {
+                    props.onToggled(e.target.checked);
+                }}
+            />
+            <label className="toggle-label" htmlFor={props.id}>
+                <span className="toggle-inner" />
+                <span className="toggle-toggle-switch" style={props.switchStyle} />
+            </label>
         </div>
-    );
+    </div>
+
+    const labeled = <div className='inputContainer padding-lr'>
+        <label className='inputLabel inputLabelUnfocused' htmlFor={props.id}>{props.label}</label>
+        {main}
+    </div>
+
+    return props.label ? labeled : main;
 }